+++ src/event/ChartEvent.js
@@ -114,8 +114,8 @@ export default class ChartEvent {
     if (this._chartData.dragGraphicMarkFlag()) {
       this._graphicMarkEventHandler.pressedMouseMoveEvent(event)
       // 这里判断一下，如果是在拖拽图形标记，让十字光标不显示
-      if (this._chartData.crosshair().paneTag) {
-        this._chartData.setCrosshairPointPaneTag(null, null)
+      if (this._chartData.crosshair().paneId) {
+        this._chartData.setCrosshairPointPaneId(null, null)
       }
     }
     if (this._checkZoomScroll()) {