+++ src/event/ZoomScrollEventHandler.js
@@ -42,7 +42,7 @@ export default class ZoomScrollEventHandler extends EventHandler {
 
   mouseLeaveEvent (event) {
     if (isMouse(event)) {
-      this._chartData.setCrosshairPointPaneTag(null, null)
+      this._chartData.setCrosshairPointPaneId(null, null)
     }
   }
 
@@ -51,9 +51,9 @@ export default class ZoomScrollEventHandler extends EventHandler {
       return
     }
     this._performCross(event, false, cross => {
-      this._chartData.setCrosshairPointPaneTag({ x: event.localX, y: cross.y }, cross.tag)
+      this._chartData.setCrosshairPointPaneId({ x: event.localX, y: cross.y }, cross.paneId)
     }, () => {
-      this._chartData.setCrosshairPointPaneTag(null, null)
+      this._chartData.setCrosshairPointPaneId(null, null)
     })
   }
 
@@ -89,7 +89,7 @@ export default class ZoomScrollEventHandler extends EventHandler {
     this._performCross(event, true, cross => {
       if (!this._touchPoint && !this._touchCancelCrossHair && !this._touchZoomed) {
         this._touchPoint = { x: event.localX, y: event.localY }
-        this._chartData.setCrosshairPointPaneTag({ x: event.localX, y: cross.y }, cross.tag)
+        this._chartData.setCrosshairPointPaneId({ x: event.localX, y: cross.y }, cross.paneId)
       }
     })
   }
@@ -106,11 +106,11 @@ export default class ZoomScrollEventHandler extends EventHandler {
         const radius = Math.sqrt(xDif * xDif + yDif * yDif)
         if (radius < 10) {
           this._touchPoint = { x: event.localX, y: event.localY }
-          this._chartData.setCrosshairPointPaneTag(crossHairPoint, cross.tag)
+          this._chartData.setCrosshairPointPaneId(crossHairPoint, cross.paneId)
         } else {
           this._touchCancelCrossHair = true
           this._touchPoint = null
-          this._chartData.setCrosshairPointPaneTag(null, null)
+          this._chartData.setCrosshairPointPaneId(null, null)
         }
       } else {
         this._touchCancelCrossHair = false
@@ -124,12 +124,12 @@ export default class ZoomScrollEventHandler extends EventHandler {
       if (isTouch(event)) {
         if (this._touchPoint) {
           this._touchPoint = { x: event.localX, y: event.localY }
-          this._chartData.setCrosshairPointPaneTag(crossHairPoint, cross.tag)
+          this._chartData.setCrosshairPointPaneId(crossHairPoint, cross.paneId)
           return
         }
       }
       const distance = event.localX - this._startScrollPoint.x
-      this._chartData.setCrosshairPointPaneTag(crossHairPoint, cross.tag)
+      this._chartData.setCrosshairPointPaneId(crossHairPoint, cross.paneId)
       this._chartData.scroll(distance)
     })
   }
@@ -137,7 +137,7 @@ export default class ZoomScrollEventHandler extends EventHandler {
   longTapEvent (event) {
     this._performCross(event, true, cross => {
       this._touchPoint = { x: event.localX, y: event.localY }
-      this._chartData.setCrosshairPointPaneTag({ x: event.localX, y: cross.y }, cross.tag)
+      this._chartData.setCrosshairPointPaneId({ x: event.localX, y: cross.y }, cross.paneId)
     })
   }
 
@@ -160,12 +160,12 @@ export default class ZoomScrollEventHandler extends EventHandler {
       return
     }
     let isPerform = false
-    for (const tag in this._paneContentSize) {
-      const size = this._paneContentSize[tag]
+    for (const paneId in this._paneContentSize) {
+      const size = this._paneContentSize[paneId]
       if (event.localY > size.contentTop && event.localY < size.contentBottom) {
         isPerform = true
         if (performFuc) {
-          performFuc({ tag, y: event.localY - size.contentTop })
+          performFuc({ paneId, y: event.localY - size.contentTop })
         }
         break
       }