+++ docs/zh-CN/CHANGELOG.md
@@ -1,5 +1,12 @@
 # 版本日志
 
+## 6.0.3
+`待发布`
++ 🆕 新增配置`candle.tooltip.defaultValue`，`technicalIndicator.tooltip.defaultValue`和`separator.activeBackgroundColor`。
++ 💄 优化拖拽调整图表高度显示。
++ 🐞 修复`candle.tooltip.values`自定义颜色问题。
+
+
 ## 6.0.2
 `2020-12-17`
 + 💄 优化x轴十字光标时间显示。